+++ integrations/plotly/utils/api_requests.py
@@ -2,7 +2,7 @@ import requests
 import json
 import os
 from datetime import datetime
-from plot_utils import plot_json_data_in_gui
+from utils.plots import plot_json_data_in_gui
 
 def fetch_and_show_data(data_file, strategy, start_ts, end_ts, multi_positions):
     url = f"http://127.0.0.1:5000/QTSBE/{data_file}/{strategy}?start_ts={start_ts}&end_ts={end_ts}&multi_positions={multi_positions}&details=True"
